AI/requirements.txt
# Core dependencies
torch==2.6.0
numpy>=1.24.0
websocket-client>=1.6.0
gymnasium>=0.29.0
tqdm>=4.66.0
matplotlib>=3.7.0
pandas>=2.0.0

# CUDA support
--extra-index-url https://download.pytorch.org/whl/cu118
